datatools/tg/db/requirements.txt
# Database dependencies for Telegram data storage
psycopg2-binary>=2.9.0
click>=8.0.0